Skip check for rcwhat link in rsyncToPhot.py

+++ bin/rsyncToPhot.py
@@ -46,6 +46,8 @@ if __name__ == '__main__':
     elif len(flphot) > len(fls):
         # remove bad links
         for fl in flphot:
+            if 'rcwhat' in fl:
+                continue
             if os.path.islink(fl):
                 if not os.path.exists(os.readlink(fl)):
                     os.unlink(fl)
